+++ structures/__init__.py
@@ -3,5 +3,5 @@
 # pylint: disable = invalid-name
 default_app_config = "structures.apps.StructuresConfig"
 
-__version__ = "2.10.0"
+__version__ = "2.11.0a1"
 __title__ = "Structures"